gtk: Put module of `list_view_apps_launcher` in separate folder
