+++ examples/list_view_apps_launcher/application_row/application_row.ui
